Consoliate state module in babylon contract
